+++ packages/md-bundler/src/index.ts
@@ -1,4 +1,2 @@
 export { bundleMdx } from './bundle-mdx'
 export { generateToc, generateTocFromHtml } from './generate-toc'
-export { getMDXComponent, getMDXExport } from 'mdx-bundler/client'
-export type { MDXContentProps } from 'mdx-bundler/client'